fix: disable additional rules handled by TypeScript compiler (#283)
Rules disabled: `no-import-assign`, `no-setter-return`, `import/no-named-as-default-member`

While this could be seen as a breaking change, in reality the TypeScript code would not compile. The code must adhere to these rules already. So the impact of this change is simply reducing unnecessary noise within linting errors (and leaving it to the TypeScript compiler, which is better suited).